refactor: Change page filenames
For better understanding

Resolves #54
